fix Staff component columns on smaller screen sizes
