--- components/Staff/StaffMap.jsx
@@ -1,18 +1,0 @@
-const StaffMap = [
-  {
-    name: 'Name',
-    title: 'Title',
-    image: '/assets/staffExampleImage.jpg',
-  },
-  {
-    name: 'Name',
-    title: 'Title',
-    image: '/assets/staffExampleImage.jpg',
-  },
-  {
-    name: 'Name',
-    title: 'Title',
-    image: '/assets/staffExampleImage.jpg',
-  },
-]
-export default StaffMap
