+++ components/Staff/StaffMap.tsx
@@ -1,0 +1,38 @@
+interface StaffMap {
+  name: string
+  title: string
+  image: string
+}
+const StaffMap = [
+  {
+    name: 'Name',
+    title: 'Title',
+    image: '/assets/staffExampleImage.jpg',
+  },
+  {
+    name: 'Name',
+    title: 'Title',
+    image: '/assets/staffExampleImage.jpg',
+  },
+  {
+    name: 'Name',
+    title: 'Title',
+    image: '/assets/staffExampleImage.jpg',
+  },
+  {
+    name: 'Name',
+    title: 'Title',
+    image: '/assets/staffExampleImage.jpg',
+  },
+  {
+    name: 'Name',
+    title: 'Title',
+    image: '/assets/staffExampleImage.jpg',
+  },
+  {
+    name: 'Name',
+    title: 'Title',
+    image: '/assets/staffExampleImage.jpg',
+  },
+]
+export default StaffMap
