+++ components/Staff/StaffMember.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image'
 
 export default function StaffMember({ name, title, image }) {
   return (
-    <div className="text-center mx-8 mb-4">
+    <div className="text-center max-w-[7rem] md:max-w-full">
       <Image
         className="rounded-full"
         src={image}
